+++ packages/terra-popup/src/terra-dev-site/test/popup/ArrowHorizontalAttachmentsPopup.test.jsx
@@ -75,7 +75,9 @@ class AlignmentPopup extends React.Component {
             className={cx('popup-button')}
             onClick={this.handleButtonClick}
             ref={this.setButtonNode}
-          />
+          >
+            {''}
+          </button>
         </div>
         <p>Choose Target Attachment:</p>
         <button type="button" id="attach-Top" value="top right" onClick={this.handleAttachment}>Attach Top Right</button>