+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -3,6 +3,8 @@ ChangeLog
 
 Unreleased
 ----------
+### Fixed
+* Fix JS error when losing focus from an empty date input. 
 
 4.10.0 - (August 14, 2019)
 ------------------