+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -6,6 +6,9 @@ Unreleased
 ### Changed
 * Replaced Object.assign syntax with Object spread syntax
 
+### Fixed
+* Fix JS error when losing focus from an empty date input. 
+
 4.10.0 - (August 14, 2019)
 ------------------
 ### Fixed