+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -3,7 +3,11 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Replaced Object.assign syntax with Object spread syntax
+
 ### Fixed
+* Fix JS error when losing focus from an empty date input. 
 * Fixed ambiguous hour that occurs in the southern hemisphere.
 
 4.10.0 - (August 14, 2019)