+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -3,6 +3,9 @@ ChangeLog
 
 Unreleased
 ----------
+
+4.11.0 - (August 21, 2019)
+------------------
 ### Changed
 * Replaced Object.assign syntax with Object spread syntax